Create Activity 23: C-Strings
From the activity description: "Write a program that asks the user to enter a series of single digit numbers with nothing separating them. Read the input as a c-string. Make sure you input no more than 25 characters.

The program should display the sum of all the single-digit numbers in the string. For example, if the user enters 25a1y4, the program should display 12, which is the sum of 2, 5, 1, and 4. Treat any non-digit character as a '0'

You may NOT use the cstring library with this program.
You may use the iomanip library to limit input."